requirements.txt
# Flask Web框架
Flask==3.1.2
Werkzeug==3.1.5

# 模板引擎
Jinja2==3.1.6
MarkupSafe==3.0.3

# YAML配置
PyYAML==6.0.3

# 浏览器自动化（PNG生成）
playwright==1.57.0

# 生产环境WSGI服务器
gunicorn==23.0.0

# 工具库
python-dateutil==2.9.0

# 环境变量管理
python-dotenv==1.0.1
